pyproject.toml
[tool.poetry]
name = "hg-twitter"
version = "0.1.0"
description = "Hieroglyph Twitter Bot"
authors = ["sali0 <saadali93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^3.9.0"
twitter-text-python = "^1.1.1"
google-cloud-storage = "^1.32.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
